Extract some of the setup functions to a separate file
